Refactor suji and kabe helpers
