+++ project/game/ai/first_version/helpers/defence.py
