+++ project/game/ai/first_version/tests/defence/tests_defence.py
@@ -1,0 +1,8 @@
+# -*- coding: utf-8 -*-
+import unittest
+
+from mahjong.tests_mixin import TestMixin
+
+
+class DefenceTestCase(unittest.TestCase, TestMixin):
+    pass
