+++ project/game/ai/first_version/tests/defence/tests_kabe.py
@@ -1,0 +1,8 @@
+# -*- coding: utf-8 -*-
+import unittest
+
+from mahjong.tests_mixin import TestMixin
+
+
+class KabeTestCase(unittest.TestCase, TestMixin):
+    pass
