+++ project/game/ai/first_version/tests/defence/tests_suji.py
@@ -1,0 +1,8 @@
+# -*- coding: utf-8 -*-
+import unittest
+
+from mahjong.tests_mixin import TestMixin
+
+
+class SujiTestCase(unittest.TestCase, TestMixin):
+    pass
